% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann2ppp.R
\name{ann2ppp}
\alias{ann2ppp}
\title{ann2ppp Read a h5ad file and convert it to a point process object from spatstat library}
\usage{
ann2ppp(
  data_path,
  mask = NULL,
  magnif,
  spat_coords_name = c("x", "y"),
  marks_col = "type"
)
}
\arguments{
\item{data_path}{path to the h5ad file}

\item{mask}{Optional: img object to be converted in the window of the point process class object}

\item{magnif}{level of magnification of the original H&E image, can be 20x or 40x}

\item{spat_coords_name}{names of coordinate of the centroids in the h5ad file}

\item{marks_col}{name of the .obs vector with the names of the marks (usually the cell types names)}
}
\value{
An object of class "ppp"
}
\description{
ann2ppp Read a h5ad file and convert it to a point process object from spatstat library
}
