% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_kde.R
\name{downsample_KDE}
\alias{downsample_KDE}
\title{downsample_KDE Function to downsample based on the Kernel Density Estimation.}
\usage{
downsample_KDE(ppp_obj, marks_name, n_target, ...)
}
\arguments{
\item{ppp_obj}{object of class "ppp" from spatstat}

\item{marks_name}{vector of dim 2 with the names of the cell types to apply the function to}

\item{n_target}{number of point in the ppp object to downsample}

\item{...}{additional specification for the density.ppp function}
}
\value{
An object of class "ppp" with n_target points for each marks_name
}
\description{
It needs a pp objects, the marks name which is a vector of dim 2 and a target number of point to which downsample.
It's possible to personalized the density.ppp function
}
