% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smoothed_lfun.R
\name{plot_smoothed_Lfun}
\alias{plot_smoothed_Lfun}
\title{plot_smoothed_Lfun Function to plot the smoothed Lcross object}
\usage{
plot_smoothed_Lfun(
  smoothed_obj,
  clusters_vec,
  mark_i = "Mark 1",
  mark_j = "Mark 2",
  center_plot = TRUE
)
}
\arguments{
\item{smoothed_obj}{list with elements "r", breaks vector, and "fd", object of the fd class}

\item{clusters_vec}{vector of cluster value for each sample}

\item{mark_i}{name of the mark value for which the Lcross function was calculated}

\item{mark_j}{name of the mark value for which the Lcross function was calculated}

\item{center_plot}{to choose if you want the plot centered (L-r ~ r) or not.}
}
\value{
A ggplot object
}
\description{
plot_smoothed_Lfun Function to plot the smoothed Lcross object
}
